add new permission to the reserve trip view
